Deleted index file.
