--- assets/js/db.py
@@ -1,18 +1,0 @@
-import psycopg2
-import sys
-from  flask import Flask,render_template
-from flask import jsonify
-
-app = Flask(__name__)
-
-@app.route('/data')
-def send_data():
-    con = psycopg2.connect("host='localhost' dbname='Non_profit' user='postgres' password='Bobo90'")  
-    cur = con.cursor()
-    cur.execute("""select * from  income""")
-    data = [col for col in cur]
-    cur.close()
-    return jsonify(data)
-
-if __name__ == "__main__":
-    app.run(debug=True)
