+++ assets/templates/index.html
@@ -1,0 +1,5 @@
+<body>
+    <h1> Python Fetch Example</h1>
+    <p id='embed'>{{embed}}</p>
+    <p id='mylog'></p>
+    <body>
